# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Assure Logistics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ShipmentStatus = typeof ShipmentStatus[keyof typeof ShipmentStatus];


export const ShipmentStatus = {
  pending: 'pending',
  in_transit: 'in_transit',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  delayed: 'delayed',
  cancelled: 'cancelled',
} as const;

export interface Shipment {
  id: number;
  trackingNumber: string;
  origin: string;
  destination: string;
  status: ShipmentStatus;
  carrierId: number;
  /** @nullable */
  carrierName?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  estimatedDelivery?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface ShipmentInput {
  /** @minLength 1 */
  trackingNumber: string;
  /** @minLength 1 */
  origin: string;
  /** @minLength 1 */
  destination: string;
  carrierId: number;
  weight?: number;
  description?: string;
  estimatedDelivery?: string;
  status?: string;
}

export interface ShipmentUpdate {
  origin?: string;
  destination?: string;
  status?: string;
  carrierId?: number;
  weight?: number;
  description?: string;
  estimatedDelivery?: string;
}

export interface ShipmentEvent {
  id: number;
  shipmentId: number;
  type: string;
  description: string;
  /** @nullable */
  location?: string | null;
  createdAt: string;
}

export interface ShipmentEventInput {
  /** @minLength 1 */
  type: string;
  /** @minLength 1 */
  description: string;
  location?: string;
}

export interface Carrier {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  activeShipments?: number;
  createdAt: string;
}

export interface CarrierInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
  contactName?: string;
  phone?: string;
  email?: string;
}

export interface CarrierUpdate {
  name?: string;
  code?: string;
  contactName?: string;
  phone?: string;
  email?: string;
}

export type DashboardSummaryStatusBreakdownItem = {
  status: string;
  count: number;
};

export interface DashboardSummary {
  totalShipments: number;
  inTransit: number;
  delivered: number;
  delayed: number;
  totalCarriers: number;
  pendingCount: number;
  statusBreakdown?: DashboardSummaryStatusBreakdownItem[];
}

export type ListShipmentsParams = {
status?: string;
carrierId?: number;
search?: string;
};