# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Assure Logistics API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: shipments
    description: Shipment management
  - name: carriers
    description: Carrier management
  - name: events
    description: Shipment events
  - name: dashboard
    description: Dashboard aggregates
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /shipments:
    get:
      operationId: listShipments
      tags: [shipments]
      summary: List shipments
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: carrierId
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shipment"
    post:
      operationId: createShipment
      tags: [shipments]
      summary: Create a shipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"

  /shipments/recent:
    get:
      operationId: listRecentShipments
      tags: [shipments]
      summary: List recent shipments for the activity feed
      responses:
        "200":
          description: Recent shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shipment"

  /shipments/{id}:
    get:
      operationId: getShipment
      tags: [shipments]
      summary: Get a shipment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "404":
          description: Not found
    patch:
      operationId: updateShipment
      tags: [shipments]
      summary: Update a shipment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
    delete:
      operationId: deleteShipment
      tags: [shipments]
      summary: Delete a shipment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /shipments/{id}/events:
    get:
      operationId: listShipmentEvents
      tags: [events]
      summary: List events for a shipment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShipmentEvent"
    post:
      operationId: createShipmentEvent
      tags: [events]
      summary: Add an event to a shipment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentEvent"

  /carriers:
    get:
      operationId: listCarriers
      tags: [carriers]
      summary: List carriers
      responses:
        "200":
          description: List of carriers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Carrier"
    post:
      operationId: createCarrier
      tags: [carriers]
      summary: Create a carrier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarrierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carrier"

  /carriers/{id}:
    get:
      operationId: getCarrier
      tags: [carriers]
      summary: Get a carrier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Carrier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carrier"
        "404":
          description: Not found
    patch:
      operationId: updateCarrier
      tags: [carriers]
      summary: Update a carrier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarrierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carrier"
    delete:
      operationId: deleteCarrier
      tags: [carriers]
      summary: Delete a carrier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Shipment:
      type: object
      required: [id, trackingNumber, origin, destination, status, carrierId, createdAt]
      properties:
        id:
          type: integer
        trackingNumber:
          type: string
        origin:
          type: string
        destination:
          type: string
        status:
          type: string
          enum: [pending, in_transit, out_for_delivery, delivered, delayed, cancelled]
        carrierId:
          type: integer
        carrierName:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    ShipmentInput:
      type: object
      required: [trackingNumber, origin, destination, carrierId]
      properties:
        trackingNumber:
          type: string
          minLength: 1
        origin:
          type: string
          minLength: 1
        destination:
          type: string
          minLength: 1
        carrierId:
          type: integer
        weight:
          type: number
        description:
          type: string
        estimatedDelivery:
          type: string
        status:
          type: string

    ShipmentUpdate:
      type: object
      properties:
        origin:
          type: string
        destination:
          type: string
        status:
          type: string
        carrierId:
          type: integer
        weight:
          type: number
        description:
          type: string
        estimatedDelivery:
          type: string

    ShipmentEvent:
      type: object
      required: [id, shipmentId, type, description, createdAt]
      properties:
        id:
          type: integer
        shipmentId:
          type: integer
        type:
          type: string
        description:
          type: string
        location:
          type: ["string", "null"]
        createdAt:
          type: string

    ShipmentEventInput:
      type: object
      required: [type, description]
      properties:
        type:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        location:
          type: string

    Carrier:
      type: object
      required: [id, name, code, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        contactName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        activeShipments:
          type: integer
        createdAt:
          type: string

    CarrierInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
        contactName:
          type: string
        phone:
          type: string
        email:
          type: string

    CarrierUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        contactName:
          type: string
        phone:
          type: string
        email:
          type: string

    DashboardSummary:
      type: object
      required: [totalShipments, inTransit, delivered, delayed, totalCarriers, pendingCount]
      properties:
        totalShipments:
          type: integer
        inTransit:
          type: integer
        delivered:
          type: integer
        delayed:
          type: integer
        totalCarriers:
          type: integer
        pendingCount:
          type: integer
        statusBreakdown:
          type: array
          items:
            type: object
            required: [status, count]
            properties:
              status:
                type: string
              count:
                type: integer