# lib/api-zod/src/generated/types/carrier.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Assure Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface Carrier {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  activeShipments?: number;
  createdAt: string;
}