# lib/api-zod/src/generated/types/carrierInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Assure Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface CarrierInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
  contactName?: string;
  phone?: string;
  email?: string;
}