# lib/api-zod/src/generated/types/carrierUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Assure Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface CarrierUpdate {
  name?: string;
  code?: string;
  contactName?: string;
  phone?: string;
  email?: string;
}