# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Assure Logistics API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryStatusBreakdownItem } from './dashboardSummaryStatusBreakdownItem';

export interface DashboardSummary {
  totalShipments: number;
  inTransit: number;
  delivered: number;
  delayed: number;
  totalCarriers: number;
  pendingCount: number;
  statusBreakdown?: DashboardSummaryStatusBreakdownItem[];
}