# lib/api-zod/src/generated/types/shipment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Assure Logistics API
 * OpenAPI spec version: 0.1.0
 */
import type { ShipmentStatus } from './shipmentStatus';

export interface Shipment {
  id: number;
  trackingNumber: string;
  origin: string;
  destination: string;
  status: ShipmentStatus;
  carrierId: number;
  /** @nullable */
  carrierName?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  estimatedDelivery?: string | null;
  createdAt: string;
  updatedAt?: string;
}