# lib/api-zod/src/generated/types/shipmentEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Assure Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface ShipmentEvent {
  id: number;
  shipmentId: number;
  type: string;
  description: string;
  /** @nullable */
  location?: string | null;
  createdAt: string;
}