# lib/api-zod/src/generated/types/shipmentEventInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Assure Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface ShipmentEventInput {
  /** @minLength 1 */
  type: string;
  /** @minLength 1 */
  description: string;
  location?: string;
}