# lib/api-zod/src/generated/types/shipmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Assure Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface ShipmentInput {
  /** @minLength 1 */
  trackingNumber: string;
  /** @minLength 1 */
  origin: string;
  /** @minLength 1 */
  destination: string;
  carrierId: number;
  weight?: number;
  description?: string;
  estimatedDelivery?: string;
  status?: string;
}