# lib/api-zod/src/generated/types/shipmentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Assure Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface ShipmentUpdate {
  origin?: string;
  destination?: string;
  status?: string;
  carrierId?: number;
  weight?: number;
  description?: string;
  estimatedDelivery?: string;
}